Merged PR 2748: Shows full NPR version number in POS status bar
It first attempts to retrieve the version number from the last known app id from app.json. If it can find that app in published app and then installed app for the tenant, it uses the information from the "Published Application" table.

If this fails, it falls back to previous logic (which retrieves it from the "NPR Upgrade History" table.

Keep in mind - it should not fail! The only possible scenario in which it fails (and uses fallback method) is when app.json changes the id of the application (which should never be done!). If anyone ever changes the app id, then the app ID should be changed in the GetNPRVersion function in codeunit 6150702.

Related work items: #1756