Case 425730: Added property UsageCategory = Administration to all pages
Case 425730: Added property UsageCategory = Administration to all pages not having the property

This is just a temporary solution, as not all pages must have the property assigned. We'll need to remove the property from pages, which shouldn't be included for users in search results in BC.
We'll need to do it gradually later on.

Related work items: #1749