Case 425479: next set of compile warnings fixed
**I'm not sure what we should do with ExternalStorageInterface.Codeunit.al:**
it designed to run an obsoleted method '**OnAttachBinaryFile**', and a replacement for the method is expected to be added in a future BC version (17.1).
I've found a comment made by a person (not MS employees): "You can use this function in 16.x. Obsoleted does not mean removed." But how it would be possible to pass MS certification, if we left the code in our solution?
So didn't find any other solution but commenting out the source code lines.
Please do let me know, if I should've taken a different approach here.

Other fixes included in the PR:
- NPGDPRManagement.Codeunit.al(354,24): warning AL0432: Field 'Picture' is marked for removal
  - Replaced obsoleted field 'Picture' by field 'Image'
- DocExchangeActivities.Page.al(155,26): warning AL0432: Method 'CreateIncomingDocumentFromServerFile' is marked for removal
  - Replaced obsoleted method by suggested alternatives
- MagentoSetupMgt.Codeunit.al(1094,24): warning AL0432: Page 'Code Coverage Object' is marked for removal
  - Replaced obsoleted page by suggested alternatives
- +Restored accidentally commented lines in EventCalendarMgt.Codeunit.al

Related work items: #1749