Merged PR 3206: Obsolete Codeunit, deleted
The functionality in this codeunit was obsoleted by Transcendence. The check will not occur since its guarded by salesperson code changing from blank to valid using validate. This kind of happens when you have locked and then unlock POS, but the from value is not blank at that point.
The problem intended to be solved is sales ticket inclusion on balancing for AR and pre-transcendence.

Codeunit deleted

Related work items: #2000