PR 3246: Case 421880: POS quote management improvements during balancing
During balancing process, if there are POS quotes in the system, 2 options are available for the user:
1) Delete all saved POS Quotes
2) Terminate/Delete saved POS Quotes manually

Once the 2nd option is selected, a list with POS quotes will be opened, and one can close it without removing any sales and it should keep all of them for the next day. However, this is not very intuitive if the user is not aware of the behaviour. Thus the option to open the list has been renamed to "Review saved POS Quotes".
Confirmation messages have been added on OK and Cancel buttons of POS quote list page:
* OK: You have not closed all the POS Quotes. If you continue with the balancing process, the remaining quotes will stay in the system.\Are you sure you want to continue?
* Cancel: Are you sure that you want to cancel the balancing process?

Related work items: #2034