Merged PR 3631: Big DotNet refactoring of front-end communication part of code
This pull request includes all files from the `topic/front-end-dotnet-refactoring` branch developed for BC160. All the content was moved manually from that branch, and file-by-file comparison was done because of the restructuring done on BC170 master branch by Alen (he needed to separate the Application code from the Test code, but this made it impossible to merge my BC160 branch because all of the files were moved and weren't available on their original paths anymore).

In any case, this PR now compiles and runs and my tests show me that everything is fine and exactly as it should be.

Related work items: #2433