Merged PR 3825: Replacing .NET in codeunit 6014578
Replacing .NET in codeunit 6014578

This PR is a replacement for https://dev.azure.com/navipartner/NpCore/_git/NpCore/pullrequest/3019 (which I am abandoning). Please check the original PR for reference.

I have also decided not to refactor MemoryStream inside of this codeunit because it's a big bowl of spaghetti that should be addressed as a module.

The primary reason I want this PR out is because of `codeunit 6014495 "NPR Text Encoding Mapper"` that I want out as it will help people on other cases.

Related work items: #1873