Merged PR 3921: Made POS entries page unsearchable (can still be found from Object list)
I changed the UsageCategory from "Administration" to "None" in POSEntries.Page.al. as the POS entries page seems to have the same info as the POSEntryList, but displayed in a much uglier and cluttered way. Additionally, POS entries also has some print problems, which is another reason to use POSEntryList instead.
---
Changing the UsageCategory to "None" makes the page unsearchable from "Tell me"(the search at the top right corner), but it is still able to be found in the object list.

Related work items: #2672