Merged PR 4168: Added PromotedOnly property to all pages with the promoted property
I wrote a script that automatically added the promotedOnly property to every action that currently has the promoted property. This was both done in order to get rid of the warning/info AW0011 and because it looks better to not have duplicated actions. (This is the second PR, the first PR had some errors).

Related work items: #2837