Merged PR 4248: Aligned the discount factbox on the itemCard
I aligned the discount factbox on the itemCard, by changing the trigger for the period discount to "OnAsssistEdit". This is because the other ones were more to the left due to the 3 dots coming from the above mentioned trigger. So adding this trigger to the last discount, added the 3 dots and thus aligned them.

Related work items: #2869