Merged PR 4380: Added promotedCategory property to promoted actions on pages (and pageExt) that didn't already have the property
Wrote a PS script that added the promotedCategory property to promoted actions on pages and pageExt where the property had not already been set. As a standard I just set the promotedCategory to process (and again of course not overriding already set promotedCategory properties).
--
There are some limitations to the script, in that if a page for example has two promoted actions but only one of them had a promotedCategory property set, the script would skip that file and not add the promotedCategory to the action missing it. I tried to do a few checks on whether or not this was this case in any page files, and it didn't seem to be and if there are some cases, there aren't a lot.

Related work items: #2935