Merged PR 4403: Created a generic retail headline that is currently used on retail admin RC
Created a generic retail headline that can be used on multiple RCs since it doesn't have any role specific info. Currently I have set it to be in use on the retail admin RC, as that one doesn't really have any specific info that makes sense to display in a headline.

Related work items: #2943