Merged PR 4473: Cleaned up legacy code.
I know that this pull request is insane.
I did consider whether or not I should take it in meaningful steps for other people to review, but once I started last week I realized how much dead code was still in NPRetail and how slow a proper process would be.
We had it all blindly converted to AL, now organized in folders, but it was still a ton of dead code intertangled with new code all over the place.
No one in charge ever did proper cleanup and thus this was code going back 20 years.
The problem for developers without historic context is they would need to confirm with a colleague or Mark constantly what is safe to delete and what isn't which slows them down a lot compared to
jumping through references in vscode and knowing yourself that "there might be 3 references but these are also dead".
Another problem with trying to clean up one module at a time would have been that all these old references are circular.

Now, I have most likely broken something unintended but it's hopefully worth it from an economic POV even though I hate to create a PR so chaotic.

The following is a non-complete list of things done in this PR:
# Obsoleted the following modules:
- Company sync (old CHANGECOMPANY sync via i-comm setup)
- Dynamic Module
- Gift Voucher
- Credit Voucher
- Foreign Gift Voucher & Credit Vouchers
- Alternative Number
- Google Cloud Print
- Hardcoded Retail Prints
- Customer Location (a hacky precursor to the restaurant module)
- External Storage
- File import Scanner functions (doesn't work in webclient)
- Warranty
- Used Goods
- Retail Document
- Table Import Export

# Other things deleted:
- Old POS Sales Document handling where you stayed in active sale while ping-ponging with the sales doc.
- Label Barcode on item card
- A bunch of v1 actions tied to dead code,
- A bunch of duplicate v1 actions. We had 4 actions for reversing a normal POS receipt, 3 actions for exporting to sales documents. With slight variations.

# Other changes:
- Refactored end of sale routine to be independent of the huge old auditroll intertangled RetailFormCode.FinishSale routine.
- Refactored start of sale to force use of a non-locking number series.

In total, deleted ~70k lines of dead code.

-----
I've asked Mark to create cases on a bunch of auxiliary problems spotted while working on this.
We can not create a new release until we've had some thorough testing of the master branch.
Until then, we will have to create hotfixes for the existing release.

Related work items: #2985