Merged PR 4512: Re-added customer no. search support
Previously, the EAN box subscriber was in the huge receivables action.
Since that was deleted, it needed to be re-added.
This has been done in the simpler customer selection action.

Some pages that shouldn't be searchable have also been cleaned up and a dead action was deleted.

Related work items: #3029