Merged PR 4673: Added actions to deploy packages from Azure Stor.
I have added actions to deploy packages from Azure Blob Storage.

Because configuration package is searched (get) by filename and becase request was that there is possibility to have multiple versions of same package, I have added logic with sufix '_ver' so if file on azure has it in its name, it will truncated so it will end up like configuration package code, for example:
   Ticket_ver2_basic.radipstart
   Ticket_ver3_advanced.rapidstart
   Ticket_ver4.rapidstart
would all be Get('Ticket"); like it was in the past from
   Ticket.rapidstart
because in the past only '.rapidstart' was truncated.

Related work items: #2903