Merged PR 4635: Created a control add in for a getting started wizard page
I created a control add In for a getting started wizard page, that is very much like Microsofts. As I am not the greatest designer, I chose to just copy Microsofts design, this can be changed in the future :). Also currently the page links to 3 sort of random NP videos because no videos to the specific topics have been made yet. So as soon as the videos have been made, I will change the URLs to the new videos.

Related work items: #3080