Merged PR 4714: Added vscode task for injecting minified javascript into actions
Added vscode task for injecting minified javascript into actions.
This is to be prepare for clean WF2 actions that no longer need to have javascript in a string.

I have tested it on some of the self-service actions with a bunch of code. It minifies everything into 1 long line, and escapes ' with '' as per AL requirements.

It could be added to a vscode extension later, but lets start with a workspace task to see it in action.

Related work items: #3139