Merged PR 4723: Adding the Open POS action on the role center
It's not possible to open a tablet client using Hyperlink from within a non-tablet session. The only way we can expose POS through role center is with the Web client.

Another problem is that AL cannot even generate the correct tablet client URL anymore, if I choose tablet, I get this: `'ms-dynamicsnav://bc:80/BC?company=CRONUS%20International%20Ltd.&page=6150750'` (that's old desktop-style URL for the RTC client!!)

I tried this with a regex `\/((default|tablet)(\.aspx)?)?\?` (leaving it here for posterity) and then use Regex.Replace to replace it with `/Tablet?` but I get into that same problem I identified couple of weeks ago from Major Tom - it starts attempting to load non-existing pages (it goes to SigninTablet page which doesn't exist).

From all this, my suspicion is that Microsoft is probably not actively doing anything on the Tablet client, and that support for it has fallen into some state of "disrepair" and that maybe we should also start thinking of actually using the regular Web client instead.

Related work items: #3147