Merged PR 4792: Finished refactoring the JSON Management codeunit
The case started as a possible issue with JSON view setup, however after a short investigation it turned out that the problem in fact was that parameter parsing from menu button setup was incorrect. Root cause was that when JSON Management fails on retrieving parameter values from action configuration in menus (or from context object during execution of workflows) it doesn't actually give any meaningful context as to where exactly and why it failed.

I've decided to refactor JSON Management codeunit and all consumers so as to provide the necessary context.

Now parameter and value retrieval functions are more explicit, they come with plain (e.g. GetStringParameter) version that never fails and returns default value for a type and strict (e.g. GetStringParameterOrFail) version that fails if the parameter/value referenced is not found. SetScope functions have been duplicated similarly except that the strict version does not return a boolean parameter and has the same conceptual logic of for example invoking Rec.Insert() without inspecting the output parameter, and plain version returns boolean with the same conceptual logic of for example invoking if Rec.Insert() then

Also, all of public methods in JSON Management have been documented with XMLDoc so that people can have and use IntelliSense on them when writing workflow code.

Related work items: #3198