Merged PR 4836: Changed the way the POS view picture is displayed on the POS View Profile card
Added a factbox with the POS view picture to the POS View profile card. I moved the actions that dealt with importing, exporting and deleting the picture to the new page. Additionally I created a maximum size for the pictures (150 kB) to ensure the POS doesn't suffer performance issues because of the size of the picture.

Related work items: #3233