Merged PR 4889: Workaround rapidstart bug on container creation
We see a problem when this codeunit is invoked where it can fail on a missing table 6014404 (the first table in npretail sorted by ID).

This does not make sense: We are invoking one object in npretail while another does "not yet exist".
This implies some race condition between publishing on NST and invoking a codeunit via powershell on NST.

This is an attempt at working around the problem by delaying at the start of the codeunit.

Related work items: #3281