Merged PR 5000: Bug fix - BC to Mag. sync using XML (JSON was OK)
Bug fix - BC to Mag. sync using XML (JSON was OK).

And a little bit of fine tune. In this case we had an error in the Data also (Magento was not completely configured) so in BC we had only "Bad Request" in the response window.
Soap UI is returning the same, but Postman was returning an error that tax_class_id is empty.
So now in the response window we will display response content together with the ReasonPhrase if HTPP request is not OK (not 200).

Related work items: #3406