Merged PR 5145: Relays not firing on subsequent ticket scans
When scanning ticket in speedgate and ticket is not valid due to error:

"[-1004] - The capacity for ENTBIL00000217314 has been exceeded. Entry is not allowed."

the error response was suppressed, and speedgate seemed unresponsive to ticket scan.
Error is now returned to speedgate, so speedgate can show a red light.

Related work items: #3577