Merged PR 5248: Re-including the pre-transcendence fix in Model UI
Re-including the pre-transcendence fix in Model UI

Adyen integration in old database that are currently being migrated to BC170 still use the `n$.Framework.RaiseObjectModelEvent` call to send protocol events from model UI to AL. That method originates in Standard and when framework was updated to Transcendence back in 2016, that call was replaced with `n$.respond` and for Transcendence people have been writing Model UI protocols that use the new `n$.respond` syntax. When providing support for Model UI in Dragonglass, the support for pre-Transcendence `n$.Framework.RaiseObjectModelEvent` was removed because there was not a single instance in AL where this was called, and all Model UI protocols consistently use the `n$.respond` (or `n$.respondExplicit`) calls.

However, it turns out that integration of Adyen in the 456726 case for some reason uses an app that still uses the old Standard-era syntax.

To avoid more issues of this kind, I am re-implementing the fix to convert the pre-Transcendence syntax into Transcendence syntax, because other than this Adyen integration that uses the old syntax, there may be other customer apps outside of NP Core that are also still using this old syntax.

Related work items: #3594