Merged PR 5263: Add currency_factor to Magento Sales order import
The purpose of this is to make sure that whatever exchange rate was used to convert the web order from base currency to another currency is used for posting the order as well.
This is to allow us to add support for multiple currencies in some of our related modules in the future.

Related work items: #3025