Merged PR 5431: MT and HW conn. Download actions URL changed.
Major Tom and Hardware Connector Download actions URL changed. We are not using Azure Blob Storage because of maintenance complexity. Instead we are forwarding user to already existing URLS for downloads.

Actions had to be moved to Cues because action on RoleCenter type page can't have triggers.

Related work items: #3308