Merged PR 5605: Fix #450630 - Standard notification bar in POS
This PR solves the problem of the notification bar shown in BC inside POS. For POS - that part of UI is simply not going to be shown.

This applies to both Transcendence and Dragonglass.

This PR also puts the previous "fresh" theme (softer blues) back, rather than using the NaviPartner website blue color as the main theme color.

Related work items: #3984