Merged PR 5636: NpXml new features and one bug fix
NpXml new features and one bug fix.

Because some previous requests and now because of Nordiska Galleriet, new features are aded to our NpXml module:

1.	Custom Namespace URI with default alias (xmlns) is now possible. I have added field “Custom Namespace for XMLNS Alias.
It does not depend on “Namespace Enabled” field because often can happen that Namespace is not needed at all. They just want custom Namespace URI for default alias.

2.	Xml attributes were possible only with child elements inside root element, but not on Root Element by itself. Now it is possible on root element also. If you enable “Root Attributes Enabled”
you will get four fields more.
I have decided for this approach because it is not easy to use standard XmlAttributes table because it is connected to XmlElement table, but there we don’t store Root Element.
Because it has a lot of references and findsets I think that it is not smart. So I have added four fields for two attributes and I doubt that we ever will need more than two.
Even if it happen, we will just add one pair of fields and two lines of code.

3.	Current Date and Time as value of some element in the file is now possible by calling the function GetCurrentDateAndTime from Codeunit 6151562:

4.	Very specific request from Nordiska, but from other perspective, it could be needed again for some other client. Round values on wanted number of decimals and to choose should it be up or down.
And even to round it as a whole number, also with precision. So in the XmlElement List I have added two field. "Round precision" and "Round direction"
This is applied for both ways of getting values (From a Decimal field) or from Custom Codeunits if returned value is decimal.

More details in the case (attached email with screenshots)

Related work items: #3987