PR 5717: Case 459852. Possibility to include parent item categories in XML templates. + more user friendly object id lookups
Possibility to include parent item categories in XML templates. Case 459852
It was not possible to extract multiple levels of parent item categories to an XML file, generated from an NP XML template. This was because of inherited filters from previous xml elements, set-up on XML Template Field Mapping, which couldn't be overridden due to filtergroup usage.

\+ more user friendly object id lookups - with object caption

Related work items: #4068, #4069