Merged PR 5874: Default Admission Per POS Unit
# **Changes to Default Admission Selection when selling and scanning tickets, supporting multi-site setup.**

When building complex multi-site and multi-admission tickets, the setup required that each site would have its own set of ticket items for the automatic process to work in regard to auto-admitting a ticket directly on sales.
There were also issues when the same POS would want to use the “Input Box” (fka “EAN Box”) to scan tickets with different capabilities and configuring 1 common admission code was not possible.

## Single site vs multi-site

As single site ticket with multiple admission could be setup with automatic admission to the default admission when the ticket was sold. This is achieved using settings on the “Ticket Type” and “Ticket BOM” accordingly:

“Ticket Type” specifies that on sales, the admission marked as default will admitted automatically. Ticket Type have multiple activation methods to choose from:
•	Scan
•	(POS) Default Admission
•	(POS) All Admissions
Selecting the “Scan” option will not admit any admission on sales and is typically used for web tickets.
Having one ticket granting access to multiple sites and automatically admitting the ticket at the local site, does not work since there is only one default admission. This was solved creating different items for different sites, configuring different defaults.

**Activation Method on Ticket BOM**

The “Activation Method” can now also be specified on the “Ticket BOM” (provided that configuration source is “Ticket BOM”).
This will allow for a higher degree of granularity when selecting which admission should be automatically selected in different scenarios.

With above setup, it will be backward compatible and admission behavior will remain unchanged.
The new activation methods on “Ticket BOM” are:
•	(blank): Initial Value – will use the setup defined on “Ticket Type”
•	On Scan: This admission will be selected when ticket is scanned from a POS Unit and the suggested admission code is blank, allowing the system to select the admission code.
•	On Sales: This admission code will be selected when a ticket is sold from the POS.
•	Always: This admission code will be selected either when you sell the ticket or when ticket is scanned and the suggested admission code is blank, allowing the system to select the admission code.
•	Per Unit: Specific rules for individual POS units can be setup. Note that each POS unit that scan tickets and the suggested admission code is blank must have a rule.

**Important**
When scanning a ticket and system is allowed to select an appropriate admission code – admission will fail when no admission is selected.

#**Multi-Site support**

Consider the following model:

Site A has 2 admissions (A1 and A2) and Site has 2 admissions (B1 and B2). When ticket is sold on POS 01 at site A, A1 should be auto admitted and when sold from POS 02 at site B, B1 should be auto admitted.
(Previously this could not be achieved on same ticket and wo...