Merged PR 5871: Supporting lookup functionality from front end
Supporting lookup functionality from front end

The lookup functionality allows front-end fields to send lookup requests to the back end and to retrieve the data relevant for user input. For now it's very relevant for balancing. For example, when looking up bank bin codes during balancing, the user must specify which bin to put the money into. This is typically achieved using lookup in BC. With this PR we can achieve this also from POS.

The objects in this PR are providing the back-end infrastructure to support lookup. It also includes PoC lookup for bank deposit bin code.

Related work items: #4035