Merged PR 6437: Notes (RecordLinks) bug in Magento Sal. Orders FIX
Notes (RecordLinks) bug in Magento Sales Orders fixed. BC was showing message "The data cannot be shown because it is not valid."

It's needed to use Codeunit "Record Link Management" when Notes are inserted. I have changed only in Magento module. Tested on the example from case 465719.

I saw there are more places in NpCore where Notes are created. I don't know if same problem exists in those functions (it is possible that is the same).

Related work items: #4278