Merged PR 6768: Sales Order import bugs Reading invalid text data
Sales Order import bugs Reading invalid text data

We had two bugs in this case. First one is because encoding. File is downloaded using Azure function in base64 and saved in stream. Encoding was not needed.
By this error from case 467733 was solved.

But going further, It came that it says that NC Task is processed but order is not created.
There was one more bug in finding XML element.

Both fixed and tested in case container.

Related work items: #4386