Merged PR 6769: 449123 DSFINVK export for Germany
Fiskaly was having trouble with DSFINVK exports, now everything is working.
At the end of work shift (Z report) we are sending json file of all POS receipts (that are send to Fiskaly API) to DSFINVK API.
We cant update existing Cash Point Closing on DSFINVK API, so GUID is saved only when DSFINVK is
successfully send (response is 200).
DSFINVK Job codeunit is created just like Fiskly, to send every record with error again, and to check if Cash Point Closing have COMPLETED status and if he does then export is trigerred.
Exports when they are triggered are visible on FIskaly dashboard.

Related work items: #4205