Merged PR 7195: Going back to blobs from media
Going back to blobs from media.
Media code commented out for now to make it easier to switch back again.

- "MM Member".Image replaced with Picture
- "NPR MM Member Info Capture".Image replaced with Picture
- "NPR AF Args: Spire Barcode".Picture replaced with Image
- "NPR MCS Faces".Image replaced with Image
- "NPR MPOS QR Code"."QR Image" replaced with "QR code"
- "NPR Display Content Lines".Picture replaced with Image (had to return piece of dotnet here, guesstimation of image extension was done via dotnet)
- "NPR Retail Logo"."POS Logo" replaced with Logo
- "NPR RP Template Media Info".Image replaced with Picture
- "NPR NpRv Arch. Voucher"."Barcode Image" replaced with Barcode
- "NPR NpRv Voucher"."Barcode Image" replaced with Barcode

Related to case: 473889

I kept new factboxes on pages where they were added. Exporting of images is lost for now as there is no info on filetype stored in blobs and i did not want to pollute code with dotnet.
If needed i'll add it, but this functionality was not present in previous version so it won't be missed for now.

Related work items: #4544, #4545