Merged PR 7228: Custom Options, Previous/Next actions error fixed
Custom Options, Previous/Next actions error fixed.

Error was displayed when user goes back on previous custom option because it will try to do insert but that already exist. In the past we had repeater on this page. When it was removed, we moved to logic with Next/Previous actions. But it keeps filter, so I have put reset.

Regarding change on  page 6151428 "NPR Magento Item Cstm Options", it is because that page was visible in search but it should not. It is card page.

Related work items: #4541