PR 7336: POS Action 'PRINT_EXCH_LABEL': refactored to WF2
POS Action 'PRINT_EXCH_LABEL' has been refactored to WF2.
Now it will use new WF2 popup object dialog types "calendarPlusLines" and "datepad".

There is an issue left: currently we cannot affect which lines are shown on the "calendarPlusLines" (or I'm not aware of such a way). System will show all lines of current sale (including the lines, for which exchange labels may not be required, such as comment lines), leaving it up to the user to decide on the lines exchange labels are to be printed for.
We may improve this in another PR.

Case 470246

Related work items: #4548