Merged PR 7390: Fill Item desc. on Mag. Ord. imp. if filled in XML
Fill Item description on Magento Order import if it is filled in the XML file.

Till now Item description was filled while creating Sales Line and Type is Item, field "No." is being validated and Description was automatically filled in from the Item card and field: Description.

Now I changed that the function "InsertSalesLineItem" so it overwrites description if it was filled in the XML file. If it is empty, it will be filled from the Item as it was till now.

Related work items: #4606