Merged PR 7558: Notification processing improvement
Notification processing improvement

If the messaging API crashes due to a wrongly constructed message, the NAS will reattempt the message indefinitely. Resulting in no messages being sent.

This change will preemptively mark the message as failed before message is attempted and update it with sent status on success.

(Worst case scenario is that _all_ messages will be marked as failed due to the problem.)

Related work items: #4663