Merged PR 7548: Fixed encoding issue when import Sales Ord. by FTP
Fixed encoding issue when import Sales Ord. by FTP

OK, here we have not so nice history of this kind of errors.
Encoding issue was fixed by adding TextEncoding parameter in PR 4978 by case 453988
but it was removed later in PR 6768 by case 467733.

Now I had to put it back and realized that it will not be good again, so I had to take my time and test it deeper. The initial cause was that Azure function which is downloading files from FTP was doing it in Base64 but without encoding. That is the reason why it was OK for API import, but not for FTP.

It is fixed now and works for both as it should.

Related work items: #4636