Merged PR 7602: Enforce NoImplicitWith
Enforce NoImplicitWith.

- AL0604 is already marked with Action = Error in main.ruleset file.
- But this does not actually enforce no implicit with. It is still a warnig.
(_We want to remove the implicit with, but also offer an opt-in model to avoid forcing everyone to upgrade their code at once.)_
[https://docs.microsoft.com/en-us/dynamics365/business-central/dev-itpro/developer/devenv-deprecating-with-statements-overview
](https://docs.microsoft.com/en-us/dynamics365/business-central/dev-itpro/developer/devenv-deprecating-with-statements-overview
)

To make it error we have 2 options:
1) Add pragmas to each page and codeunit: #pragma implicitwith disable
2)  Set the NoImplicitWith flag in app.json.

I think the 2nd option is better for us as we want the warning for all objects. Enabling this will also make it easier to remove when it will not be needed: I think MS will enforce this and maybe #pragma or NoImplicitWith will not be needed anymore in the future.

Related work items: #4703