Merged PR 7674: Preventing failure in data binding when incorrect data source and setup is pa...
Preventing failure in data binding when incorrect data source and setup is passed to front end

This is mostly defensive. There was case 465797 which has incorrectly configured the sale view: the view explicitly referenced captionbox fields that were excluded from the data source by a switch in the back end. This has caused a failure in data binding.

Still, there should be no runtime error due to incorrect data binding, and instead of crashing, the renderer should render whatever is correct, and indicate in the UI and the log what's incorrect.

That's what this change takes care of.

Related work items: #4722