Merged PR 7859: Replenishment Transfer Mapping (Mag. Sal. O. imp.)
**This is new feature in NP Core**. It was initiated by Aveny, but decided to go in core because it is very reasonable and nice feature to have.

It's a bit complicated to follow because it has multiple conditions and setups because of those conditions but at the end it is nice.

Replenishment Transfer Mapping when importing Magento Sales Order.

**How it works?**

If on the client's webshop, buyer make an order of some product with higher amount than available in that location (location is configured on magento website), first it will check if feature is enabled.
If this new feature is:
- not enabled: It will work as it was working till now. It will create an order with requested quantity for that
location although that location does not have it.
- enabled: It will go further. First it will search for "Replenishment Transfer Mapping" This is new page
where user can configure from which locations items will be transfered (Transfer Order will be created)
when qty. in main one is lower than requested.
If this mapping is empty but "Auto Create Req. Lines" is enabled, it will not give an error, and go further with creation of Requisition Worksheet lines.
If this mapping is empty but "Auto Create Req. Lines" is not enabled, user will receive an error that it has to insert at least one Replenishment Transfer Mapping or to enable Auto Create Req. Lines.

When it comes to quantities from locations in Transfer Mapping, it will transfer as much as it is available, and then it will go to next location from mapping (if exist) and take the rest from there. Of course, there can be as many mappings/locations as user needs.

At the end if there is no enough pcs in locations from mapping, requisition lines will be created (of course if "Auto Create Req. Lines" is enabled on Magento Setup).

This feature also supports an option for "Projected" stocks. User can enable "Include Projected Quantities" and set desired within period in the field "Projected. Qty. Within Period" and based on that, all incoming and outgoing transfers will be included in calculation. (Purchase Orders, Sales Orders, Transfer Orders, Req. worksheets...)

And feature also includes Item Variants, so if Magento Sales Order comes with item number ending with "_someVariantCode" it will be processed accordingly.

Initially tested by myself and later by Nasser.

Related work items: #4712