Merged PR 7845: Fix warning AA0139 + fix replication counter issue + add more apis
**Details about Fix Replication Counter Issue** (for code review - codeunit 6014626 "NPR Replication Counter Mgmt." and history)

Because of Microsoft limitation related to using SQL Timestamp for table extensions (full story here: https://github.com/microsoft/ALAppExtensions/issues/13474), we had to make some tricks to copy latest SQL Timestamp in our field "**Replication Counter**":

The best **logic** I could think of for now was this:
- We use OnBeforeModify and we always have Replication Counter lower than the one in the database, but still the newest one..
- For example in database we have 9. We insert(OnAfterInsert) and onBeforeModify it system puts it 9. Then because OnModify runs, in database we will have 10.. Next time when Modify runs, we will get value 10 and in database it will be 11. But 10 will still be the greatest of our Replication Counters.
- So we use OnAfterInsert, OnAfterRename, but OnBeforeModify to fill in replication counter.

The idea is that we will always be with one behind in our Replication Counter compared to the actual SQL Timestamp.

However, while testing more, today I found an **issue with this approach**.

Consider the following **scenario**:

1) HQ Company:
- Record A is modified. Actual SQL Timestamp = 11 --> Replication Counter = 10 (previous SQL Timestamp).
- Record B is inserted. Actual SQL Timestamp = 15 --> Replication Counter = 14 (previous SQL Timestamp).

2) Child company(where we replicate data to):
Above modifications are replicated to another database. Replication counter = 14 (in Replication Endpoint Setup).

3) HQ Company:
- Record A is modified again. Actual SQL Timestamp = 16 --> Replication Counter = 11 (previous SQL Timestamp of this record).

4) Child company(where we replicate data to):
At this moment we loose the last change as the replication counter was already set as 14, so system will search for records with replication counter > 14 in HQ company.

Only **solution** I can think of at the moment is that:
- when setting the replication counter for a record, like we do at step 3 above, to check if there is already a greater replication counter in that table on another record.
- if there is, we need to set the replication counter for the current record to a value greater than that one.
- because we use "IF NOT RecRefNewerRepCounter.IsEmpty()" and RecRefNewerRepCounter.FindLast(); with a index on field "Replication Counter" this should not add big overhead.
- it is not perfect solution, but is best I can think of at the moment.. Maybe we can brainstorm about how to improve it. If someone has a better idea, please let me know.

Related work items: #4775