PR 8172: Fixed discount coupon application on web
It was not possible to apply discount coupons on web.
After debugging it was discovered that the discount coupon application routine just failed, and the original request content data was returned back to the web caller, without any kind of discount calculation.

What system actually does in the background while validating received coupons - it tries to create a mock a POS sale transaction:
- new POS sale is created,
- coupons are applied to the sale,
- then system collects all the information regarding calculated discounts (as a result of the coupon application) and stores it in temporary tables,
- the POS sale transaction is rolled back.

The problem was it that we actually were not in a POS environment: not all the required objects were correctly initialized.
This has been fixed.

Case 475431

Related work items: #4821