Merged PR 8176: Custom Search Box
Custom Search Box on ticket and members list page.

This customer search box uses filtergroup -1 to do a cross column search. Only the most logical fields are included and together with domain knowledge casting of data can be kept to a minimum.

Under specific circumstances it outperforms MS Search:
* A large data set (>100 000 records)
* There will be no search on potential table extensions - the join is being prevented

It is also noticeable, that a regular filter will be faster. With this approach the full result set must be processed before a result is shown to user.
* avoid including columns having the same value on a great number of records.

UI it works well in web client, and OK on a phone. The tablet mode however, seems to be forgotten by MS (BC17). It looks like crap.

Related work items: #4825